Correcting error when trying to access the feed of a 'sub-channel' when he is tied to a channel called 'videos'.
The problem is that with the url adding 'OPPS_MULTIMEDIAS_VIDEO_CHANNEL' before the 'channel_ long_slug' regex, makes it not sent parent channel to the view. If exists a channel named 'videos' will give error because the view it will not find the sub-channel with with 'long_slug' containing the parent channel named 'videos'.